combine shell_handler and cmd_handler
