+++ Cargo.toml
@@ -1,25 +1,24 @@
 [workspace]
 members = [
+    "ex3_ground_station/cli_ground_station",
+    "ex3_ground_station/dashboard",
+    "ex3_ground_station/server",
+    "ex3_obc_fsw/bulk_msg_dispatcher",
     "ex3_obc_fsw/dev_tools/cli_test_msg",
     "ex3_obc_fsw/dev_tools/ipc_burst_hardcoded",
     "ex3_obc_fsw/dev_tools/ipc_dummy_client",
     "ex3_obc_fsw/dev_tools/ipc_server_dev_tool",
-    "ex3_obc_fsw/handlers/dfgm_handler",
+    "ex3_obc_fsw/handlers/adcs_handler",
     "ex3_obc_fsw/handlers/coms_handler",
+    "ex3_obc_fsw/handlers/dfgm_handler",
     "ex3_obc_fsw/handlers/dummy_handler",
     "ex3_obc_fsw/handlers/iris_handler",
-    "ex3_obc_fsw/handlers/cmd_handler",
     "ex3_obc_fsw/handlers/shell_handler",
     "ex3_obc_fsw/scheduler",
-    "ex3_obc_fsw/bulk_msg_dispatcher",
-    "ex3_ground_station/cli_ground_station",
-    "ex3_ground_station/dashboard",
-    "ex3_ground_station/server",
     "ex3_shared_libs/bulk_msg_slicing",
     "ex3_shared_libs/common",
-    "ex3_shared_libs/tcp_interface",
-    "ex3_shared_libs/message_structure",
     "ex3_shared_libs/interfaces/ipc",
     "ex3_shared_libs/logging",
-    "ex3_obc_fsw/handlers/adcs_handler",
+    "ex3_shared_libs/message_structure",
+    "ex3_shared_libs/tcp_interface",
 ]
