--- ex3_obc_fsw/handlers/cmd_handler/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "cmd_handler"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-message_structure = { path = "../../../ex3_shared_libs/message_structure" }
-common = { path = "../../../ex3_shared_libs/common" }
-ipc = { path = "../../../ex3_shared_libs/interfaces/ipc" }
-tcp_interface = { path = "../../../ex3_shared_libs/tcp_interface" }
-logging = { path = "../../../ex3_shared_libs/logging" }
-log = "0.4.22"
