+++ ex3_obc_fsw/handlers/shell_handler/Cargo.toml
@@ -6,3 +6,6 @@ edition = "2021"
 [dependencies]
 common = {path = "../../../ex3_shared_libs/common"}
 ipc = { path = "../../../ex3_shared_libs/interfaces/ipc" }
+log = "0.4.22"
+logging = { path = "../../../ex3_shared_libs/logging" }
+message_structure = { path = "../../../ex3_shared_libs/message_structure" }
