+++ ex3_obc_fsw/msg_dispatcher/connection.h
@@ -27,7 +27,6 @@ enum ComponentId
     GS = 7,
     COMS = 8,
     BULK_MSG_DISPATCHER = 9,
-    CMD = 10,
     SHELL = 12,
     TEST = 99,
 };
